refactor(daw-mixer): co-locate each island in its own folder with test and CSS
