--- components/client/DawMixer/DawMixer.module.css
@@ -1,158 +1,0 @@
-/* ── DawMixer client islands ────────────────────────────────────────────── */
-
-/* VuMeter — horizontal 12-segment bar, 14px height */
-.vuMeter {
-  position: relative;
-  height: 14px;
-  border: 1px solid var(--ds-color-signal-quiet);
-  background: rgba(0, 0, 0, 0.7);
-  display: flex;
-  align-items: stretch;
-  gap: 1px;
-  padding: 1px;
-  cursor: ew-resize;
-  outline: none;
-}
-.vuMeter:focus-visible {
-  outline: 2px solid var(--ds-color-signal);
-  outline-offset: 2px;
-}
-.vuSegFilled {
-  flex: 1;
-  background: var(--ds-color-signal);
-  height: 100%;
-}
-.vuSegEmpty {
-  flex: 1;
-  background: var(--ds-color-signal);
-  height: 100%;
-  opacity: 0.55;
-}
-.vuSegRed {
-  flex: 1;
-  background: var(--ds-color-feedback-error);
-  height: 100%;
-}
-
-/* FaderIsland — 18px track with center rail + 14×22px thumb */
-.faderTrack {
-  position: relative;
-  height: 18px;
-  background: color-mix(in srgb, var(--ds-color-signal) 3%, transparent);
-  border: 1px solid var(--ds-color-signal-quiet);
-  cursor: ew-resize;
-  outline: none;
-}
-.faderTrack::before {
-  content: "";
-  position: absolute;
-  left: 8px;
-  right: 8px;
-  top: 50%;
-  height: 1px;
-  background: var(--ds-color-signal-quiet);
-  transform: translateY(-0.5px);
-  pointer-events: none;
-}
-.faderTrack:focus-visible {
-  outline: 2px solid var(--ds-color-signal);
-  outline-offset: 2px;
-}
-.faderThumb {
-  position: absolute;
-  top: -2px;
-  left: 0;
-  width: 14px;
-  height: 22px;
-  background: var(--ds-color-signal);
-  box-shadow: 0 0 8px var(--ds-color-signal);
-  border: 1px solid #000;
-  cursor: ew-resize;
-  will-change: transform;
-}
-.faderThumb::after {
-  content: "";
-  position: absolute;
-  left: 50%;
-  top: 3px;
-  bottom: 3px;
-  width: 1px;
-  background: #000;
-  transform: translateX(-50%);
-}
-
-/* KnobIsland — 26px circle knob */
-.knob {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 3px;
-  outline: none;
-  user-select: none;
-}
-.knob:focus-visible {
-  outline: 2px solid var(--ds-color-signal);
-  outline-offset: 2px;
-}
-.knobSvg {
-  cursor: ns-resize;
-}
-.knobLabel {
-  font-size: var(--ds-font-size-2xs);
-  color: var(--ds-color-text-muted);
-  text-align: center;
-  letter-spacing: 0.1em;
-  font-family: var(--ds-font-family-mono);
-}
-
-/* RmsButtons — 18×18px squares */
-.rmsButtons {
-  display: flex;
-  gap: 3px;
-}
-.rmsActive {
-  min-width: 24px;
-  min-height: 24px;
-  padding: 0 6px;
-  border: 1px solid var(--ds-color-signal);
-  background: color-mix(in srgb, var(--ds-color-signal) 10%, transparent);
-  box-shadow: 0 0 5px color-mix(in srgb, var(--ds-color-signal) 35%, transparent);
-  color: var(--ds-color-signal);
-  font-size: var(--ds-font-size-2xs);
-  font-family: var(--ds-font-family-mono);
-  font-weight: 700;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  letter-spacing: 0.02em;
-  flex-shrink: 0;
-}
-.rmsInactive {
-  min-width: 24px;
-  min-height: 24px;
-  padding: 0 6px;
-  border: 1px solid var(--ds-color-signal-quiet);
-  background: transparent;
-  color: var(--ds-color-text-muted);
-  font-size: var(--ds-font-size-2xs);
-  font-family: var(--ds-font-family-mono);
-  font-weight: 700;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  letter-spacing: 0.02em;
-  flex-shrink: 0;
-}
-.rmsActive:focus-visible,
-.rmsInactive:focus-visible {
-  outline: 2px solid var(--ds-color-signal);
-  outline-offset: 2px;
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .faderThumb {
-    box-shadow: none;
-  }
-}
