+++ components/client/DawMixer/FaderIsland/FaderIsland.test.tsx
