+++ components/client/DawMixer/KnobIsland/KnobIsland.module.css
@@ -1,0 +1,22 @@
+.knob {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 3px;
+  outline: none;
+  user-select: none;
+}
+.knob:focus-visible {
+  outline: 2px solid var(--ds-color-signal);
+  outline-offset: 2px;
+}
+.knobSvg {
+  cursor: ns-resize;
+}
+.knobLabel {
+  font-size: var(--ds-font-size-2xs);
+  color: var(--ds-color-text-muted);
+  text-align: center;
+  letter-spacing: 0.1em;
+  font-family: var(--ds-font-family-mono);
+}
